crew/pyproject.toml
[project]
name = "pipeline-safety-rag-crew"
version = "0.1.0"
description = "Multi-agent CrewAI crew for PHMSA pipeline safety regulations — feature branch of gas-and-energy-mechanics-copilot"
authors = [{ name = "Ashish Gupta", email = "ashishanilgupta@gmail.com" }]
requires-python = ">=3.11,<3.14"
dependencies = [
    "crewai[tools]>=0.108.0",
    "faiss-cpu>=1.9.0",
    "boto3>=1.35.0",
    "pandas>=2.0.0",
    "pyarrow>=18.0.0",
    "numpy>=1.26.0",
]

[project.scripts]
run_crew = "pipeline_safety_rag_crew.main:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline_safety_rag_crew"]

[tool.crewai]
type = "crew"
